payment-service/requirements.txt
flask==3.1.1
flask-cors==5.0.1
gunicorn==23.0.0
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf>=5.28.2,<6.0
